requirements.txt
SPARQLWrapper
pandas
pytest
rdflib-jsonld
requests
